file structure refactor
