+++ README.md
@@ -51,6 +51,31 @@ We use `pre-commit` with `ruff` to check formatting and linting before commits a
 pre-commit run --all-files
 ```
 
+## Project Directories
+
+```bash
+.
+├── .github/                          # GitHub workflows for CI/CD tests
+├── configs/                          # Training and evaluation configuration YAMLs
+├── experiments/                      # Scripts and setups for running controlled experiments
+├── results/                          # Output from experiments: plots, logs, metrics, etc.
+├── scripts/                          # Shell scripts for launching jobs (e.g. on WatGPU)
+├── simglucose/                       # Custom fork of simglucose package
+├── src/                              # Main RL framework package
+│   ├── agents/                       # Code for creating, loading, and managing agents
+│   ├── environments/				  # Custom Gym environments and reward functions
+│   ├── training/                     # Orchestration of training, prediction, and evaluation
+│   ├── utils/                        # Generic helper utilities
+│   ├── __init__.py
+│   └── main.py                       # Entry point for training/evaluation
+├── tests/                            # Unit and integration tests
+├── .gitignore
+├── .pre-commit-config.yaml           # Pre-commit checks configuration
+├── README.md
+├── requirements.txt                  # Python package dependencies
+└── setup.py                          # Install script to register `src` as a module
+```
+
 ## Managing the `simglucose` Fork
 
 This repository includes `simglucose` as a git subtree. If you need to update or push changes to the [simglucose fork](https://github.com/Blood-Glucose-Control/simglucose.git), use the following commands **in the root directory**.
